refactor: ChatRoomDetail 컴포넌트 Container/Presentational 패턴으로 분리
